Create a way to provide MediaInfo for dynamic DataSources to ShadowMediaPlayer.
Before SDK version 24 there was no way to directly specify an
AssetFileDescriptor as the data source. As a result, clients of MediaPlayer has
to provide AssetFileDescriptor.getFileDescriptor() to setDataSource().
Now, opening an AssetFileDescriptor for the same resource multiple times can result in different internal FileDescriptors whose toString() don't return the same
String. However, ShadowMediaPlayer uses FileDescriptor.toString() to find a match which results in an InvalidArgumentException due to missing MediaInfo.

This change allows tests to pass a provider of MediaInfo which will be used for creating MediaInfo for any DataSource.

PiperOrigin-RevId: 326514181